created an npm seed command to make seeding database easier. moved the few routes that were in server.js into their appropriate controllers/routes folders too
